Removing FCM token route made and implemented. Notifications security upgraded.

+++ frontend/src/components/TopNavbar.jsx
@@ -4,15 +4,27 @@ import { FaUser } from "react-icons/fa";
 import Cookies from "js-cookie";
 import { useNavigate } from "react-router-dom";
 import { FaRobot } from "react-icons/fa";
+import axios from "axios";
+import toast from "react-hot-toast";
 
 const TopNavbar = () => {
   const navigate = useNavigate();
 
   const handleSignOut = () => {
+    const userId = Cookies.get("id");
+    const response = axios.post("https://fairfare-0hyl.onrender.com/user/remove-fcm-token", {
+      userId: userId,
+    });
+
+    if(response.status === 200){
     Cookies.remove("id");
     Cookies.remove("last4");
     localStorage.clear();
     navigate("/");
+    }
+    else{
+      toast.error("Error signing out. Please try again.");
+    }
   };
 
   return (
